Initial commit
This JMH suite compares the speed of contains() for ArrayList<Object>
and HashSet<Object>. It includes a few large collections to sanity-check
our expectation of Big O-guarantees and many small collections to
examine those guarantees in the context of CPU caches and prefetching.

Anecdotally, the amount of "small" collections greatly outnumber the
amount of "large" collections. Big O

    > describes the limiting behavior of a function when the argument
    > tends towards a particular value or infinity [1]

but is occasionally misleading for small arguments (here collections),
especially in the face of hardware advances over the past two decades.

[1] https://en.wikipedia.org/wiki/Big_O_notation