Prove no hash code collision
To prove that the difference between the throughput of
Collection<Object>::contains and Collection<String>::contains is not
because of hash code collisions, allocate 10k objects and count the
unique hash codes.

This says nothing about where the difference then comes from, only that
it must be something else. I suspect String::equals -- that accounts for
the similar pattern for both ArrayList and HashSet, and we know that
String::equals is worst-case O(n) in String::length.